+++ packages/plugin-analytic-fb-pixel/src/main.ts
@@ -1,39 +1,6 @@
 import { pluginLoader } from '@ima/core';
 
-import {
-  FacebookPixelAnalytic,
-  type AnalyticFBPixelSettings,
-} from './FacebookPixelAnalytic';
-
-declare global {
-  interface Window {
-    fbq: facebook.Pixel.Event;
-    _fbq: facebook.Pixel.Event;
-  }
-}
-
-export interface PluginAnalyticFBPixelSettings {
-  fbPixel: {
-    id: string | null;
-  };
-}
-
-declare module '@ima/core' {
-  // eslint-disable-next-line @typescript-eslint/no-empty-interface
-  interface PluginAnalyticSettings extends PluginAnalyticFBPixelSettings {}
-
-  interface PluginSettings {
-    analytic: PluginAnalyticSettings;
-  }
-
-  interface Settings {
-    plugin: PluginSettings;
-  }
-
-  interface OCAliasMap {
-    '$Settings.plugin.analytic.fbPixel': AnalyticFBPixelSettings;
-  }
-}
+import { FacebookPixelAnalytic, type AnalyticFBPixelSettings } from './FacebookPixelAnalytic';
 
 pluginLoader.register('@ima/plugin-analytic-google', () => ({
   initSettings: () => ({
@@ -49,6 +16,5 @@ pluginLoader.register('@ima/plugin-analytic-google', () => ({
   }),
 }));
 
-export { FacebookPixelAnalytic };
-
-export type { AnalyticFBPixelSettings };
+export type { PluginAnalyticFBPixelSettings } from './types';
+export { FacebookPixelAnalytic, type AnalyticFBPixelSettings };