+++ packages/plugin-analytic-fb-pixel/src/types.ts
@@ -1,0 +1,31 @@
+import type { AnalyticFBPixelSettings } from './FacebookPixelAnalytic';
+
+declare global {
+  interface Window {
+    fbq: facebook.Pixel.Event;
+    _fbq: facebook.Pixel.Event;
+  }
+}
+
+export interface PluginAnalyticFBPixelSettings {
+  fbPixel: {
+    id: string | null;
+  };
+}
+
+declare module '@ima/core' {
+  // eslint-disable-next-line @typescript-eslint/no-empty-interface
+  interface PluginAnalyticSettings extends PluginAnalyticFBPixelSettings {}
+
+  interface PluginSettings {
+    analytic: PluginAnalyticSettings;
+  }
+
+  interface Settings {
+    plugin: PluginSettings;
+  }
+
+  interface OCAliasMap {
+    '$Settings.plugin.analytic.fbPixel': AnalyticFBPixelSettings;
+  }
+}