+++ packages/plugin-analytic-google/src/main.ts
@@ -1,37 +1,6 @@
-/* eslint-disable @typescript-eslint/no-empty-interface */
 import { pluginLoader } from '@ima/core';
 
-import {
-  GoogleAnalytics4,
-  type AnalyticGoogleSettings,
-} from './GoogleAnalytics4';
-
-declare global {
-  interface Window {
-    gtag: Gtag.Gtag;
-    dataLayer: unknown[];
-  }
-}
-
-export interface PluginAnalyticGoogleSettings {
-  google4: AnalyticGoogleSettings;
-}
-
-declare module '@ima/core' {
-  interface PluginAnalyticSettings extends PluginAnalyticGoogleSettings {}
-
-  interface PluginSettings {
-    analytic: PluginAnalyticSettings;
-  }
-
-  interface Settings {
-    plugin: PluginSettings;
-  }
-
-  interface OCAliasMap {
-    '$Settings.plugin.analytic.google4': AnalyticGoogleSettings;
-  }
-}
+import { GoogleAnalytics4, type AnalyticGoogleSettings } from './GoogleAnalytics4.js';
 
 pluginLoader.register('@ima/plugin-analytic-google', () => ({
   initSettings: () => ({
@@ -53,6 +22,5 @@ pluginLoader.register('@ima/plugin-analytic-google', () => ({
   }),
 }));
 
-export { GoogleAnalytics4 };
-
-export type { AnalyticGoogleSettings };
+export type { PluginAnalyticGoogleSettings } from './types';
+export { GoogleAnalytics4, type AnalyticGoogleSettings };