+++ packages/plugin-http-client/src/types.ts
@@ -1,0 +1,15 @@
+import type { RestResourceSettings } from './rest/AbstractResource';
+
+export interface PluginHttpClientSettings {
+  rest?: RestResourceSettings;
+}
+
+declare module '@ima/core' {
+  interface PluginSettings {
+    httpClient: PluginHttpClientSettings;
+  }
+
+  interface Settings {
+    plugin: PluginSettings;
+  }
+}