+++ packages/plugin-local-storage/src/types.ts
@@ -1,0 +1,9 @@
+import type LocalStorage from './LocalStorage';
+
+declare module '@ima/core' {
+  interface Utils {
+    LocalStorage: LocalStorage;
+  }
+}
+
+export {};