+++ packages/plugin-analytic-fb-pixel/CHANGELOG.md
@@ -1,5 +1,23 @@
 # Change Log
 
+## 6.0.6
+
+### Patch Changes
+
+- 8f5d01a: Fixed build issues with client server config
+
+## 6.0.5
+
+### Patch Changes
+
+- e2eeb2b: Turbo, eslint update, types exports, new tsconfig
+
+## 6.0.4
+
+### Patch Changes
+
+- 6d92917: Code refactorings, moved type extensions to separate file to cleanup main.ts
+
 ## 6.0.3
 
 ### Patch Changes