+++ packages/plugin-atoms/CHANGELOG.md
@@ -1,5 +1,23 @@
 <a name="5.0.1"></a>
 
+## 10.1.4
+
+### Patch Changes
+
+- 8f5d01a: Fixed build issues with client server config
+
+## 10.1.3
+
+### Patch Changes
+
+- e2eeb2b: Turbo, eslint update, types exports, new tsconfig
+
+## 10.1.2
+
+### Patch Changes
+
+- 6d92917: Code refactorings, moved type extensions to separate file to cleanup main.ts
+
 ## 10.1.1
 
 ### Patch Changes